refactor(monitor): 目录监控包化拆分与清理
- app/monitor.py 拆分为 app/monitor/ 六模块（watcher/syslimits/snapshot/dispatcher/poller/monitor），公开导入路径不变
- 删除无引用的 MonitorChain（A11）与失效的 /proc inotify 探测（A6）
- inotify 上限比较改用目录数量口径（A7）
- 监控配置先筛选后去重，避免非监控配置顶掉监控配置并对重复项告警（A9）
- 兼容模式/网络存储跳过启动期目录遍历（A10）
- stop() 不再关闭共享缓存后端（A4）
