+++ app/monitor/__init__.py
@@ -1,0 +1,14 @@
+"""
+目录监控包。
+
+- watcher.py     本地目录监控线程（watchfiles）
+- syslimits.py   系统限制探测与监控模式决策
+- snapshot.py    远程快照存取与比对
+- dispatcher.py  监控事件到整理链的分发
+- poller.py      远程目录轮询监控
+- monitor.py     Monitor 门面：装配、生命周期与健康检查
+"""
+from app.monitor.watcher import DirectoryChangeEvent, LocalDirectoryWatcher
+from app.monitor.monitor import Monitor
+
+__all__ = ["DirectoryChangeEvent", "LocalDirectoryWatcher", "Monitor"]
